backend/requirements.txt
fastapi
uvicorn
supabase
anthropic
requests
python-dotenv
apscheduler
langgraph
pillow
reportlab
weasyprint
pytest
google-genai
